+++ khoros/utils/tests/test_settings.py
@@ -80,3 +80,21 @@ def test_invalid_node_type_exception():
     node_type = resources.get_structure_collection('board')
     with pytest.raises(exceptions.InvalidNodeTypeError):
         khoros_object.settings.get_node_setting('custom.pretend_setting', 'fake-node', node_type)
+
+
+def test_sso_status_retrieval():
+    """This function tests to confirm that the :py:meth:`khoros.core.Khoros.Community.sso_enabled` method
+       is able to return a Boolean value.
+
+       .. versionadded:: 5.0.0
+    """
+    if not resources.local_test_config_exists() or not resources.local_helper_exists():
+        pytest.skip("skipping local-only tests")
+
+    # Instantiate the Khoros object
+    set_package_path()
+    khoros_object = resources.instantiate_with_local_helper()
+
+    # Test the method
+    sso_enabled = khoros_object.communities.sso_enabled()
+    assert sso_enabled is not None